FEAT: LectureResponseDto 추가
- Lecture 엔티티로부터 DTO를 생성하는 정적 메서드 from 추가
- dto폴더 안에 lecture 패키지로 이동
